project/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
Werkzeug==2.3.7
SQLAlchemy==2.0.21
WTForms==3.0.1
email-validator==2.0.0
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
Pillow>=10.0.1
python-dateutil==2.8.2
gunicorn==21.2.0
PyMySQL==1.1.0
mysql-connector-python==9.0.0

